# apps/backend/openapi.yaml
openapi: "3.1.0"
info:
  title: Spillorama Bingo API
  version: "0.1.0"
  description: |
    Backend API for the Spillorama bingo platform.

    **Authentication**: All player and admin endpoints (except `/api/auth/login`,
    `/api/auth/register`, and Swedbank webhooks) require a Bearer token in the
    `Authorization` header obtained from the login or register endpoints.

    **Error shape**: All errors return `{ ok: false, error: { code, message } }`.
    Success responses return `{ ok: true, data: <payload> }`.
servers:
  - url: http://localhost:4000
    description: Local development
  - url: https://api.spillorama.no
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  schemas:
    ApiSuccess:
      type: object
      required: [ok, data]
      properties:
        ok:
          type: boolean
          enum: [true]
        data:
          description: Endpoint-specific payload
    ApiError:
      type: object
      required: [ok, error]
      properties:
        ok:
          type: boolean
          enum: [false]
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: INVALID_INPUT
            message:
              type: string
              example: E-post er påkrevd.
...
    SwedbankIntent:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        walletId:
          type: string
        amountMajor:
          type: number
        status:
          type: string
          enum: [PENDING, PAID, FAILED, CANCELLED]
        paymentUrl:
          type: [string, "null"]
        createdAt:
          type: string
          format: date-time
...
security:
  - BearerAuth: []
paths:
...
  /api/payments/swedbank/topup-intent:
    post:
      tags: [Payments]
      summary: Create a Swedbank Pay top-up payment intent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount]
              properties:
                amount:
                  type: number
                  minimum: 1
      responses:
        "200":
          description: Payment intent with redirect URL
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiSuccess"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/SwedbankIntent"
  /api/payments/swedbank/confirm:
    post:
      tags: [Payments]
      summary: Confirm / reconcile a Swedbank Pay intent after redirect
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [intentId]
              properties:
                intentId:
                  type: string
      responses:
        "200":
          description: Updated intent
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiSuccess"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/SwedbankIntent"
  /api/payments/swedbank/intents/{intentId}:
    get:
      tags: [Payments]
      summary: Get a payment intent by ID
      parameters:
        - in: path
          name: intentId
          required: true
          schema:
            type: string
        - in: query
          name: refresh
          schema:
            type: boolean
            default: false
          description: If true, re-reconcile with Swedbank before returning
      responses:
        "200":
          description: Intent
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiSuccess"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/SwedbankIntent"
  /api/payments/swedbank/callback:
    post:
      tags: [Payments]
      summary: Swedbank Pay server-side callback (webhook — HMAC-verified)
      description: |
        **BIN-603:** HMAC-SHA256 signature verification (no Bearer auth).

        The handler computes `HMAC-SHA256(rawBody, SWEDBANK_WEBHOOK_SECRET)`
        and compares constant-time against the digest in
        `X-Swedbank-Signature`. `sha256=<hex>` and bare `<hex>` are both
        accepted; the digest is 64 hex chars (SHA-256 output).

        If the server is mis-configured (no secret), the endpoint
        returns 503 — fail-closed so ops notices immediately rather
        than silently accepting unsigned callbacks.

        Defense-in-depth: even with a valid signature, `processCallback`
        still fetches the authoritative payment status from Swedbank API
        before crediting the wallet. Neither layer can be bypassed alone.
      security: []
      parameters:
        - in: header
          name: X-Swedbank-Signature
          required: true
          schema:
            type: string
            pattern: "^(sha256=)?[0-9a-fA-F]{64}$"
          description: |
            HMAC-SHA256 digest (hex) of the raw request body, optionally
            prefixed with `sha256=`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Raw Swedbank Pay callback payload
      responses:
        "200":
          description: Acknowledged — signature verified, callback processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        "401":
          description: INVALID_SIGNATURE — signature missing, malformed, or does not match
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "503":
          description: WEBHOOK_NOT_CONFIGURED — SWEDBANK_WEBHOOK_SECRET is not set on the server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # ── Admin — Auth ──────────────────────────────────────────────────────────
...